+++ finagle-thriftmux/src/main/scala/com/twitter/finagle/ThriftMux.scala
@@ -17,7 +17,7 @@ import com.twitter.finagle.param.{
   Tracer => _,
   _
 }
-import com.twitter.finagle.server.{StackBasedServer, StackServer}
+import com.twitter.finagle.server.{BackupRequest, StackBasedServer, StackServer}
 import com.twitter.finagle.service._
 import com.twitter.finagle.stats.{
   ClientStatsReceiver,
@@ -130,6 +130,9 @@ object ThriftMux
       Mux.server.stack
       .insertBefore(StackServer.Role.preparer, Server.ServerToReqRepPreparer)
       .replace(StackServer.Role.preparer, Server.ExnHandler)
+      // this filter adds tracing annotations and as such must come after trace initialization.
+      // however, mux removes the `TraceInitializerFilter` as it happens in the mux codec.
+      .prepend(BackupRequest.traceAnnotationModule[mux.Request, mux.Response])
 
   /**
    * Base [[com.twitter.finagle.Stack.Params]] for ThriftMux servers.